refactor: move all middlewares in handlers sub package
