+++ pkg/webhooks/handlers/admission.go
@@ -8,10 +8,7 @@ import (
 	"time"
 
 	"github.com/go-logr/logr"
-	"github.com/kyverno/kyverno/pkg/config"
 	"github.com/kyverno/kyverno/pkg/tracing"
-	admissionutils "github.com/kyverno/kyverno/pkg/utils/admission"
-	jsonutils "github.com/kyverno/kyverno/pkg/utils/json"
 	"go.opentelemetry.io/otel/attribute"
 	admissionv1 "k8s.io/api/admission/v1"
 )
@@ -88,29 +85,5 @@ func Admission(logger logr.Logger, inner AdmissionHandler) http.HandlerFunc {
 		} else {
 			logger.V(4).Info("admission review request processed", "time", time.Since(startTime).String())
 		}
-	}
-}
-
-func Filter(c config.Configuration, inner AdmissionHandler) AdmissionHandler {
-	return func(logger logr.Logger, request *admissionv1.AdmissionRequest, startTime time.Time) *admissionv1.AdmissionResponse {
-		if c.ToFilter(request.Kind.Kind, request.Namespace, request.Name) {
-			return nil
-		}
-		return inner(logger, request, startTime)
-	}
-}
-
-func Verify() AdmissionHandler {
-	return func(logger logr.Logger, request *admissionv1.AdmissionRequest, startTime time.Time) *admissionv1.AdmissionResponse {
-		if request.Name != "kyverno-health" || request.Namespace != config.KyvernoNamespace() {
-			return admissionutils.ResponseSuccess()
-		}
-		patch := jsonutils.NewPatchOperation("/metadata/annotations/"+"kyverno.io~1last-request-time", "replace", time.Now().Format(time.RFC3339))
-		bytes, err := patch.ToPatchBytes()
-		if err != nil {
-			logger.Error(err, "failed to build patch bytes")
-			return admissionutils.ResponseFailure(err.Error())
-		}
-		return admissionutils.ResponseSuccessWithPatch(bytes)
 	}
 }
